backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.19.0",
    "alembic==1.12.1",
    "asyncpg==0.29.0",
    "fastapi[all]==0.104.1",
    "httpx==0.25.2",
    "loguru==0.7.2",
    "passlib[bcrypt]==1.7.4",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "sqlalchemy==2.0.23",
    "uvicorn[standard]==0.24.0",
]
